+++ spec/integration/sequence_and_naming_i3s.rb
@@ -1,4 +1,5 @@
 I3S_DEPENDENCIES = {
+  OSVolume: [],
   PlanScripts: []
 }.freeze